backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
bcrypt==4.3.0
channels==4.2.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
python-dotenv==1.1.1
email_validator==2.2.0
fastapi==0.115.14
fastapi-cli==0.0.7
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jinxed==1.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.18
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyNaCl==1.5.0
python-multipart==0.0.20
redis==6.2.0
requests==2.32.4
djangorestframework-simplejwt
rich==14.0.0
semantic-version==2.10.0
sqlparse==0.5.3
starlette==0.46.2
termcolor==2.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
psycopg2-binary
playwright
transformers==4.55.2
accelerate
sentencepiece
gunicorn